+++ vue/components/ui/atoms/icon/icon.vue
@@ -97,7 +97,7 @@ export const Icon = {
                     // if there is a custom context defined at the root that
                     // contains the item, then uses it
                     if (iconContext) {
-                        resource = iconContext(`./${this.icon}.svg?raw`);
+                        resource = this.loadFromContext(iconContext, this.icon);
                     }
                     // otherwise fallback to default strategy for the retrieval
                     // of icons, using the `try` and `catch` strategy
@@ -140,6 +140,17 @@ export const Icon = {
             this.$nextTick(() => {
                 attrs.forEach(attr => this.setSvgAttribute(attr.key, attr.value));
             });
+        },
+        loadFromContext(context, icon) {
+            // supports both webpack 5 and webpack 4 assets bundling logic
+            try {
+                return context(`./${icon}.svg?raw`);
+            } catch {
+                // defaults to webpack 4 asset bundling logic if v5
+                // retrieval resulted in an error, making it possible
+                // to use external assets loaded by raw-loader
+                return context(`./${icon}.svg`).default;
+            }
         }
     },
     mounted: function() {